requirements.txt
streamlit==1.31.0
pandas==2.1.4
numpy==1.26.3
scipy==1.12.0
openai==1.12.0
python-dotenv==1.0.0
streamlit-javascript==0.1.5
user-agents==2.2.0
mixpanel==4.10.0